requirements.txt
sphinx
sphinx-autobuild
sympy
ipython
symengine
gmpy2
matplotlib
python-flint
